% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nibabynames.R
\docType{data}
\name{nibabynames}
\alias{nibabynames}
\title{Northern Ireland baby names}
\format{
A data frame with 20500 observations on 6 variables.
  \tabular{rll}{
    [, 1] \tab year  \tab A year (1996-2019) \cr
    [, 2] \tab sex   \tab \dQuote{M} = Male; \dQuote{F} = Female \cr
    [, 3] \tab name  \tab A name \cr
    [, 4] \tab n     \tab Count of name within year and sex \cr
    [, 5] \tab rank  \tab Rank of name within year and sex \cr
    [, 6] \tab nation  \tab Nation of data source \cr
  }
}
\source{
Northern Ireland Statistics and Research Agency. 2021. \href{https://www.nisra.gov.uk/publications/baby-names-2018}{\dQuote{Baby Names Statistics Boys and Girls}}.
}
\usage{
nibabynames
}
\description{
Full baby name data from 1997 to 2018 for Northern Ireland
from the Northern Ireland Statistics and Research Agency.
}
\details{
The data are released by the Northern Ireland Statistics and Research Agency under the \href{http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/}{Open Government License v3.0}.
}
\keyword{datasets}
