% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankings.R
\docType{data}
\name{rankings}
\alias{rankings}
\title{England & Wales top-100 baby names by year}
\format{
A data frame with 1,900 observations on 4 variables.
\describe{
\item{name}{A name.}
\item{year}{A year (1904-1994). Rankings are based upon births within that decade.}
\item{rank}{Rank of the name within decade.}
\item{sex}{Sex, \code{M} for male and \code{F} for female.}
}
}
\source{
Office of National Statistics. 2016. \href{https://www.ons.gov.uk/peoplepopulationandcommunity/birthsdeathsandmarriages/livebirths/datasets/babynamesenglandandwalestop100babynameshistoricaldata}{Top 100 Baby Names Historical Data}.
}
\usage{
rankings
}
\description{
A longitudinal dataset containing the top-100 baby girl and top-100 baby
boy names over the period 1904 to 1994.
}
\details{
The data are released by the Office of National Statistics under the \href{http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/}{Open Government License v3.0}.
}
\seealso{
\link{ewbabynames}
}
\keyword{datasets}
