% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nibabynames.R
\docType{data}
\name{nibabynames}
\alias{nibabynames}
\title{Northern Ireland baby names}
\format{
A data frame with 20,500 observations on 6 variables.
\describe{
\item{year}{A year (1997-2018).}
\item{sex}{Sex, \code{M} for male and \code{F} for female.}
\item{name}{A name.}
\item{n}{Count of name within year and sex.}
\item{rank}{Rank of name within year and sex.}
\item{nation}{Nation of data source: \verb{Northern Ireland}.}
}
}
\source{
Northern Ireland Statistics and Research Agency. 2021. \href{https://www.nisra.gov.uk/publications/baby-names-2018}{Baby Names Statistics Boys and Girls}.
}
\usage{
nibabynames
}
\description{
Full baby name data from 1997 to 2018 for Northern Ireland
from the Northern Ireland Statistics and Research Agency.
}
\examples{
library(ggplot2)
library(dplyr)

nibabynames \%>\%
  mutate(Frequency = n/nrow(nibabynames)) \%>\%
  filter(name == "Joseph") \%>\%
  ggplot(aes(x = year, y = Frequency)) +
  labs(title = "Frequency of the Name Joseph in Northern Ireland", subtitle = "Over Time",
       x = "Year") +
  theme_bw() +
  geom_line()
}
\keyword{datasets}
