% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scotbabynames.R
\docType{data}
\name{scotbabynames}
\alias{scotbabynames}
\title{Scotland baby names}
\format{
A data frame with 248420 observations on 5 variables.
  \tabular{rll}{
    [, 1] \tab year  \tab A year (1996-2019) \cr
    [, 2] \tab sex   \tab \dQuote{M} = Male; \dQuote{F} = Female \cr
    [, 3] \tab name  \tab A name \cr
    [, 4] \tab n     \tab Count of name within year and sex \cr
    [, 5] \tab rank  \tab Rank of name within year and sex \cr
    [, 6] \tab nation  \tab Nation of data source \cr
  }
}
\source{
National Records of Scotland. 2021. \href{https://www.nrscotland.gov.uk/statistics-and-data/statistics/statistics-by-theme/vital-events/names/babies-first-names/babies-first-names-2020}{\dQuote{Baby Names Statistics Boys and Girls}}.
}
\usage{
scotbabynames
}
\description{
Full baby name data from 1974 to 2020 for Scotland
from the National Records of Scotland.
}
\details{
The data are released by the National Records of Scotland under the \href{http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/}{Open Government License v3.0}.
}
\keyword{datasets}
