% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ukbabynames}
\alias{ukbabynames}
\alias{ukbabynames-package}
\title{UK baby names}
\format{A data frame with 254889 observations on 5 variables.
  \tabular{rll}{
    [, 1] \tab year  \tab A year (1996-2017) \cr
    [, 2] \tab sex   \tab \dQuote{M} = Male; \dQuote{F} = Female \cr
    [, 3] \tab name  \tab A name \cr
    [, 4] \tab n     \tab Count of name within year and sex \cr
    [, 5] \tab rank  \tab Rank of name within year and sex \cr
  }}
\source{
Office of National Statistics. 2019. \href{https://www.ons.gov.uk/peoplepopulationandcommunity/birthsdeathsandmarriages/livebirths/datasets/babynamesenglandandwalesbabynamesstatisticsgirls}{\dQuote{Baby Names Statistics Girls}}.
Office of National Statistics. 2019. \href{https://www.ons.gov.uk/peoplepopulationandcommunity/birthsdeathsandmarriages/livebirths/datasets/babynamesenglandandwalesbabynamesstatisticsboys}{\dQuote{Baby Names Statistics Boys}}.
}
\usage{
ukbabynames
}
\description{
Full baby name 1996-2017 for the United Kingdom from the Office of National Statistics
}
\details{
The data are released by the Office of National Statistics under the \href{http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/}{Open Government License v3.0}.
}
\seealso{
\code{\link{rankings}}
}
\keyword{datasets}
